Create separate Redstone config section
